moved all to generic
