+++ errors/err_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	errors "github.com/projectdiscovery/utils/errors"
-	"github.com/stretchr/testify/require"
 )
 
 func TestErrorEqual(t *testing.T) {
@@ -71,39 +70,5 @@ func TestErrorCallback(t *testing.T) {
 
 	if !callbackExecuted {
 		t.Errorf("error callback failed to execute")
-	}
-}
-
-func TestEqualsAllInt(t *testing.T) {
-	testCases := []struct {
-		Base     int
-		All      []int
-		Expected bool
-	}{
-		{5, []int{5, 5, 5, 5}, true},
-		{5, []int{1, 2, 3, 4}, false},
-		{0, []int{}, true},
-	}
-
-	for _, tc := range testCases {
-		actual := errors.EqualsAll(tc.Base, tc.All...)
-		require.Equal(t, tc.Expected, actual)
-	}
-}
-
-func TestEqualsAllString(t *testing.T) {
-	testCases := []struct {
-		Base     string
-		All      []string
-		Expected bool
-	}{
-		{"test", []string{"test", "test", "test", "test"}, true},
-		{"test", []string{"test", "test1", "test2", "test3"}, false},
-		{"", []string{}, true},
-	}
-
-	for _, tc := range testCases {
-		actual := errors.EqualsAll(tc.Base, tc.All...)
-		require.Equal(t, tc.Expected, actual)
 	}
 }
