+++ generic/generic_test.go
@@ -41,3 +41,37 @@ func TestEqualsAnyString(t *testing.T) {
 		require.Equal(t, tc.Expected, actual)
 	}
 }
+
+func TestEqualsAllInt(t *testing.T) {
+	testCases := []struct {
+		Base     int
+		All      []int
+		Expected bool
+	}{
+		{5, []int{5, 5, 5, 5}, true},
+		{5, []int{1, 2, 3, 4}, false},
+		{0, []int{}, true},
+	}
+
+	for _, tc := range testCases {
+		actual := EqualsAll(tc.Base, tc.All...)
+		require.Equal(t, tc.Expected, actual)
+	}
+}
+
+func TestEqualsAllString(t *testing.T) {
+	testCases := []struct {
+		Base     string
+		All      []string
+		Expected bool
+	}{
+		{"test", []string{"test", "test", "test", "test"}, true},
+		{"test", []string{"test", "test1", "test2", "test3"}, false},
+		{"", []string{}, true},
+	}
+
+	for _, tc := range testCases {
+		actual := EqualsAll(tc.Base, tc.All...)
+		require.Equal(t, tc.Expected, actual)
+	}
+}
